requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
jinja2==3.1.4
torch==2.3.1
torchvision==0.18.1
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.3.0
pyotp==2.9.0
qrcode==7.4.2
scikit-learn==1.5.0
matplotlib==3.9.0
aiofiles==23.2.1
python-jose==3.3.0
supabase==2.5.0
python-dotenv==1.0.1
--extra-index-url https://download.pytorch.org/whl/cpu
